[v14] Explain template variables wherever they appear (#33977)
* Explain template variables wherever they appear

Closes #13377

Wherever a page in the docs mentions the `internal` or `external`
template variables as they apply to Teleport roles, link to the Teleport
Access Controls Reference
(docs/pages/access-controls/reference.mdx#roles).

A separate change (#32696) will expand the Access Controls Reference to
include more information on the `external` and `internal` template
variables.

* Respond to zmb3 feedback

Refer to `internal` and `external` as traits instead of template
variables.

* Apply suggestions from code review

Co-authored-by: Gus Luxton <gus@goteleport.com>

* Linter fixes

---------

Co-authored-by: Gus Luxton <gus@goteleport.com>