Pin Python version in Github actions - Teleterm build fix (#2523)
* Pin Python version in Github actions

The Python version used in our Github Actions workflow for Mac builds was downgraded to 3.11.

* Adjust Python version in GitHub Actions workflow

The .github/workflows/build-mac-universal.yaml file was updated to specify the necessary Python version for the workflow. Python 3.11 is currently needed until an update to gyp v0.16.1 is implemented. This change is made in response to a related issue found on the node-gyp GitHub repository.

---------

Co-authored-by: Reed Loden <reed@goteleport.com>

Source-commit: gravitational/teleport.e@b5776429c65aac9e8e06c9ccb3d9d7aa39db734e
Source-pull-requests: gravitational/teleport.e#2523