Enable optimistic locking support for users
Converts the auth client to use the new usersv1.Service instead
of the auth.ProtoService to manage users. Since the new service
leverages conditional updates this allows optimistic locking to
now be used. The legacy RPC service now calls into the new users
service to consolidate logic for all but the UpdateUser method,
which was left alone to prevent altering the semantics of update
for legacy clients. Additionally, the auth HTTP routes to get
users were removed since they don't look to be consumed and have
long been deprecated in favor of the gRPC API.