AWS OIDC IdP Configure script: remove region (#34036)
We are asking for a region when configuring the AWS IAM OIDC IdP.
The region is actually irrelevant for the configuration: IAM resources
are global.
This can cause some confusion for the user:
- is the integration only for this region?

Our recommended way to run this script is in AWS CloudShell which has an
env var (AWS_REGION) that the Go AWS SDK uses to decide which region to
use.

If the user is running this script elsehwere, where they don't have the
env var set, an error message will be shown:
... please set the AWS_REGION evironment variable.

Which should be helpful enough for them to fix the issue.