Applies RBAC predicates to user resources
This change aims to treat Teleport users like fully-fledged
resources in the RBAC system, specifically by allowing admins
to specifiy `where` clauses in Roles that apply to to the user
being created or modifies, as opposed to just the acive user
that is performing the operation.

| Operation      | `where` exptression `resource` value           |
| ---------------|------------------------------------------------|
| `CreateUser()` | The user to be created                         |
| `UpdateUser()` | The user to be modified                        |
| `DeleteUser()` | The user to be deleted                         |
| `UpsertUser()` | The user to be created (for the `Create` verb) |
|                | The user to be modified (for the `Update` verb |

Updating a user poses a potential security risk, given that our
RBAC predecate language currently has no way to reference an
"indirect object"; that is, in a `where` clause we cannot supply
both the value of a resource as it is and to the value it will
become after the update. To get around this, when updating a user,
the where clause *only* applies to the value in storage and NOT to
the value that the user will take after the update.

This leads to some inconsistencies - for example, it is possible
to create a user that meets the `Create` permissions, but then
immediately update it to a value that violates those `Create`
rules. I couldn't think of a way to prevent this that would also
not be too cumbersome to use in practice.

----
Changelog: Added RBAC role  `where` clause support for rules governing Users.