Connect My Computer: Implement in-app flow after deep link click (#33964)
* Simplify ClusterAdd

This moves ClusterAdd away from using the presentational & container
component approach. I didn't touch ClusterLogin as it's a much more
complex component.

* Simplify prop passing for ClusterConnect

Instead of passing the same set of props around, with types repeated
over and over again, let's define a single data structure that can
be passed around.

* Get rid of ts-expect-error in WorkspacesService tests

Unnecessary function mocks also got removed.

* Move getTestSetup outside of describe block

This makes it reusable in other describe blocks.

* Return isAtDesiredWorkspace from setActiveWorkspace

This will be needed for the in-app flow of the deep link. The deep link
might trigger login to a completely fresh cluster, so we must be able
to tell whether the login succeeded or not.

* Add prefill support to ClusterConnect dialog

This will be needed to show login modal with prefilled cluster address
and username when someone opens a Connect My Computer link to a cluster
to which they're not logged in yet.

* Replace openClusterConnectDialog with openRegularDialog

* setActiveWorkspace: Accept prefill values

This will be used to provide prefilled cluster address and username from
the deep link.

* Implement in-app flow after deep link click