Add AccessGraph service to user roles and ACLs
This commit adds a new Service, AccessGraph, and integrates it into user roles and ACLs. It extends the rule set types to include the "access_graph" verb and adds AccessGraph as a new resource in userACL. This will allow users to gain authorized access to the AccessGraph service. A new constant, KindAccessGraph, is also defined for better identifiability.