oci: Optimize OCI image size and build time
Use `RUN --mount=target-/ctx` to mount the Docker build context to
access the `.deb` file to install instead of `COPY`ing it in. This saves
space on the image.

Separate the installation of the Teleport `.deb` file to come after
installing the prerequisites with `apt-get`. This allows that layer to
be cached so we can build the OSS and Enterprise images sharing that
base. This reduces the network traffic too and the risk of failure due
to transient network errors. The build time of a subsequent image is
seconds so we don't lose anything due to the removal of potential
parallelism.