Installer Scripts: use bash instead of sh (#34039)
* Installer Scripts: use bash instead of sh

The agentless version of this script uses `bash`.
Using `bash` over `sh` should help the maitenance of the script.
It's being used for the agentless version and the node join script
`sh` is sometimes an alias to `bash`, and testing this script in those
systems might lead to assume that the script is working fine, even tho
when running in a real `sh` it will fail.

* use the same expression that is being used in agentless