ci: Switch linux workflows to OIDC-based AWS roles (#2551)
Switch the release-linux and build-linux roles to use the `build` 
environment so it can use the vars that specify the roles that can be 
assumed via OIDC. This gets rid of using secrets and using the `staging` 
environment.

Make the environment name parameterizable from the top level so we can 
switch to `build-staging` for development and testing of the workflows. 
This environment is not completely ready yet, but we'll be ready to use it
when it is ready.

Consolidate the `build-dep` and `build-rpm` input parameters to the
`release-linux` workflow into a single `build-os-packages` parameter as 
we've hit the limit of 10 input parameters for `workflow_dispatch` 
workflows. The `build-linux` workflow keeps them separate as there is no 
such limit for `workflow_call` events.

While we're touching this, use a hash for
`aws-actions/configure-aws-credentials` instead of a floating version as
that is what we do everywhere else for this action. This is likely
because we can pass secrets to it, even though we don't when using OIDC.
It is safer and protects our secrets from rogue versions.

Pull-Request: https://github.com/gravitational/teleport.e/pull/2551

Source-commit: gravitational/teleport.e@38808435ccaadfc6399cba47e7b14a69c349c7ca
Source-pull-requests: gravitational/teleport.e#2551