Render the error if we didn't get a disconnect reason
With the prior commit, we successfully made server-side disconnects
look less like errors, but we also made some real errors (like
connection time outs) look less like errors.

Now, we render an error indicator when we have an unexpected
disconnect. We also add a special check for "RDP connection failed"
which is used to indicate that the connection was never established
succesfully.

This also fixes an ancient bug where our strings had an extra
unprintable byte at the end. This is because the TDP codec was
interpreting the "severity" byte as part of the string, rather
than reading only the length that was specified in the message.