Allow resources to be migrated to optimistic locking with dynamo backend (#34272)
After upgrading Auth to include support for revisions when using
dynamo as the backend, updates to any pre-existing resources would
result in an error. The conditional expression provided to the
dynamo query would always fail because the item did not yet
contain a Revision attribute. We can achieve the same conditional
update by altering the conditional expression to assert that the
revision attribute does not exist if the item being updated does
not contain a revision.