Update bloat check cache key
The old key was not correctly evaluating the sha of the before
commit resulting in reusing the action cache key "Linux-" for all
builds. This resulted in continual failures in the event that the
previous build was larger than the current build until the cache
was manually cleared. The key also did not include the job name
which makes it harder to distinguish what action is consuming
the cache.

The new key references the bloat check job name and attempts to
ensure the cache key include the previous commit sha by expanding
the value and outputting it in a separate step.