Filter dangerous environment variables before reexec (#34177) (#34275)
* Filter dangerous environment variables before reexec

This change filters potentially dangerous environment variables that could result in code execution.

This seemed safest to integrate as a new struct within the `environment.go` in utils.  This struct allows us to validate variables as they are built, making sure that regardless of the source we have validated them against our filter list.

The environment specific logic was significant enough that the current and new logic was refactored into a new package `envutils`.

* Allow the easy addition of execution environment into SafeEnv

In addition this commit adds in a check to look for duplicate keys which may be attempting to overload our set values.

* Apply PR Feedback and remove env duplicate handling

* Apply additional PR feedback