ci: Add workflows for releasing legacy OCIs (#2563)
Add a pair of release/build workflows for building the legacy OCIs for
OSS, Enterprise and FIPS. The legacy OCIs are the Debian-based OCIs (as
opposed to the distroless OCIs) and will not be available from Teleport
v15 and onward. We need to maintain this workflow as long as Teleport
v14 is supported.

Of note is that this PR changes how the multi-architecture images are
constructed. The Drone pipeline built the amd64, arm64 and arm images
separately and combined them with the `docker manifest` command. This
was the original way to do this. This workflow uses `docker buildx` to
build the multi-architecture images all in one step using the
`--platform` flag.

This makes a difference because with the old way, the individual
architecture-specific images were pushed with a `-<arch>` label suffix.
These label suffixes are retained with the new way but they just point
at the same multi-architecture image that is built by `docker buildx`.

When a user does a `docker pull` of the architecture-specific images, it
will pull the part that is appropriate to their host architecture.
Previously it is possible they could pull a foreign image. This is
rarely useful, but can still be done with `docker pull --platform <foreign-arch>`
on the new images.

Source-commit: gravitational/teleport.e@7e7ce479d8078c31099e6130568b4a157f2d68fe
Source-pull-requests: gravitational/teleport.e#2563