[v14] Store original roles with user login state. (#34356)
* Store original roles with user login state.

The user's original static roles are now stored along with the user login
state. This is necessary to properly recalculate user access on access list
membership change, as SSO users will expire after 24 hours, making it
impossible to refer to the original user definition to determine the new
role membership for the user.

* Tweak comments, code formatting, remove unnecessary nil vs. empty array distinction.