[v13] desktop access: soften the disconnect error (#34422)
* desktop access: soften the disconnect error

If the Windows host (RDP server) decides to terminate the connection,
it often sends a message indicating why the disconnect is coming.
One example of a server-initiated disconnect would be when the user
clicks "log off" in the RDP session or when they restart the remote
machine.

We forward this message to the UI, but we put it in a big red ERROR
box that makes it look like something went wrong.

Errors should be reserved for unexpected disconnections, so covert
the dialog to "info" level and soften the messaging.

* Render the error if we didn't get a disconnect reason

With the prior commit, we successfully made server-side disconnects
look less like errors, but we also made some real errors (like
connection time outs) look less like errors.

Now, we render an error indicator when we have an unexpected
disconnect. We also add a special check for "RDP connection failed"
which is used to indicate that the connection was never established
succesfully.

This also fixes an ancient bug where our strings had an extra
unprintable byte at the end. This is because the TDP codec was
interpreting the "severity" byte as part of the string, rather
than reading only the length that was specified in the message.