ci: Add workflows for releasing legacy AMIs (#2562)
Add a pair of release/build workflows for building the legacy AMIs for
OSS, Enterprise and FIPS. The legacy AMIs are the non-hardened ones and
will not be available from Teleport v15 and onwards. We need to maintain
this workflow as long as Teleport v14 is supported.

These workflows cannot currently run in the new `build-stage` GitHub
Actions environment as the `teleport-stage` account has not been set up
for packer to build AMIs. We can do this later if we need to.

Source-commit: gravitational/teleport.e@76620e138d1be97ebe65e9d8d768e85694db46a7
Source-pull-requests: gravitational/teleport.e#2562