Allow user locks to affect access list membership. (#34354)
* Allow locks to deny access to access lists.

Access list membership will now be impacted by active user locks. If a user
is locked, they will not be considered a part of an access list. This, in turn
will be used for things like Okta assignments to ensure that Okta access can be
rescinded while a lock is active.

* Access list membership checker is its own struct now.

* Rebase and fix tests.