Add access graph support to Auth server (#2647)
* Add access graph support to Auth server

This code adds access graph support to the Auth server by initializing and watching an access graph service when the server starts up. It feeds the service with all the necessary teleport resources like users, nodes, roles and access requests, ensuring the service stays up-to-date with the current state of resources in the backend of the auth server.

* Add detailed comments to access graph functions

This update includes comments for each function related to the access graph. It provides a better understanding of the implemented features, such as querying the access graph, watching the auth server for events, handling web assets, and sending crucial data to the access graph service.

* Refactor logging and function parameters in access graph integration

This commit includes updates to the logging functions in the access graph integration to use shorter, more succinct methods. It also eliminates the 'log' parameter from the 'accessGraphClient' function to simplify its interface. Additional comments have been added for clarity and improved understanding of functionality.

* Fix linter issue

* Subscribe in access graph for the auth events before sending all events to solve the race condition
Remove `waitForConnectionReady` as it's not needed anymore

* Address code review comments

* Fix race condition
Add test

* GCI imports
Update TAG test to better simulate real time usage

* Linter fixes

* Move access graph integration into a new package

* Refactor accessgraph code and improve error handling

The access graph service configuration has been refactored and moved into a constant. Also, the 'Send' method in watch handling has been wrapped with 'trace' to improve error tracking and debugging.

* Fix comment

Source-commit: gravitational/teleport.e@6f579ba02979938177aa4417d818be9deacb290f
Source-pull-requests: gravitational/teleport.e#2647