Proxy TAG requests through Auth (#2663)
* Integrate Access Graph feature and optimize services

This commit entails the integration of the Access Graph feature into the existing service configuration. This feature is manifested through an 'IsEnabled' method that checks the state of the access graph. Additionally, the auth service logic now includes a flag for enabling access graph. The services have been further optimized with adjustments geared towards improving performance and readability of the code.

Proxy TAG requests through Auth

* Move TAG enabled flag from proxy to user setting/auth context

* Swap a pointer to AccessGraph config with a value.

* Add tests for Access Graph service.

The commit introduces a number of tests for the Access Graph service inside the new 'service_test.go' file. It verifies the response of the service on different scenarios, ensuring the functionality and security of the Access Graph service is working as intended.

* Implement Access Graph feature across various files

The Access Graph feature has been incorporated in 'module.go', 'authservice.proto', 'auth.go', and 'plugin.go' files. A method 'EnableAccessGraph' is also introduced in 'module.go' to allow the usage of this feature. Descriptions of Access Graph have been adjusted across files to explicitly state the feature's enablement.

* Blindly added change base on the flag expert comment

* Remove unused access graph feature code

This commit removes all references and uses of the access graph feature from the codebase. It also updates corresponding tests to reflect this removal. It also adds EnableAccessGraph() method to testing module to fix test compilation.

* Correct typos and update error message

This commit corrects a typo in a comment and updates an error message in the access graph service file. It also rearranges and refines the access check test in the service test file for better readability and organization.

* Improve error handling for access graph service

This commit improves the error handling in the access graph service by specifically checking if the access graph is not enabled in the Auth server. If so, it returns a 404 status instead of a 500 status, providing more accurate error information.

Source-commit: gravitational/teleport.e@d66162a346753855b2256bc5d8d78abb68334ef6
Source-pull-requests: gravitational/teleport.e#2663