[v14] Fixed tag builds using commit instead of tag ref (#34675)
Tag builds for Windows are broken in the following ways:

* The ref used to determine asset version includes a short commit hash
  (this will cause promotion issues)

* Upon retry of a build (or a new tag with the same commit), builds will
  fail as the old release is not properly cleaned up

Changing the reference used to a branch/tag reference should fix the
issue for both. Note that the second issue is still bugged for push
pipelines, but we almost never retry them so it's not really worth
fixing there right now.

Co-authored-by: fheinecke <23390735+fheinecke@users.noreply.github.com>