[v14] Prepare `SearchBar` for the integration with unified resources (#34543)
* Change styling of the filter buttons, add a close icon

This matches the latest design.

* Allow selecting multiple resource filters

This will be needed for unified resources integration,
where it is possible to select more than one filter.

* Use the same filter types as unified resources

Otherwise, we will have to convert the resource filter type to the type
expected by the unified resources component and vice-versa.
It is just easier to use the same type.

* Center picker container horizontally, don't allow it to be too narrow

* Remove `undefined` from `searchResources.filters`

* Move `SUPPORTED_RESOURCE_TYPES` to const

* Make resource filter keys unique

* Remove max width from the filter buttons

* Extract `SupportedResourceType` from the unified resource type

* Rename `SupportedResourceType` to `ResourceTypeFilter`