[TAG] Add localStorage value for enabling TAG & SQL editor (#34111)
* Add localStorage value for enabling TAG

* Add key for enabling the manual SQL editor

* Add access graph features retrieval in UserService

The code adds a `fetchAccessGraphFeatures` in UserService that retrieves and stores the access graph features from the provided API URL into the local storage. It also includes error handling to log the error and continue if fetching access graph features fails. The changes also touch the config file to include the API URL for fetching the access graph features.

---------

Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>