[v14] Add user login state original traits and indicator that original values set. (#34712)
* Add user login state original traits and indicator that original values set.

Original traits has been added to the user login state for use by a watcher in
the enterprise repo that will react to user access changes. This field will be
used to rebuild the last known good state when user access changes.

Additionally, an internal annotation that indicates the original role/trait
fields have been set in the user login state has been introduced so that the
watcher knows whether the user login state can be used by this watcher.

* GCI.

* Tweak tests.

* Further tweak tests.

* Formatting.

* Fix missing argument.

* Standardize delete comment.