[v14] TAG changes (#34733)
* Add TAG UI (#34570)

* Add TAG UI

* Add missing license

* Remove elkjs

* Proxy TAG requests through Auth (#34642)

* Proxy TAG requests through Auth

* Integrate Access Graph feature and optimize services

This commit entails the integration of the Access Graph feature into the existing service configuration. This feature is manifested through an 'IsEnabled' method that checks the state of the access graph. Additionally, the auth service logic now includes a flag for enabling access graph. The services have been further optimized with adjustments geared towards improving performance and readability of the code.

* Move TAG enabled flag from proxy to user setting/auth context

* Swap a pointer to AccessGraph config with a value.

* Add 'TestNewAccessGraph' function to 'useracl_test.go' file.

The function 'TestNewAccessGraph' has been added to run tests checking the configuration of Access Graph. Furthermore, a small change is made for the 'KindAccessGraph' comment in 'constants.go' to provide more accurate information.

* Implement Access Graph feature across various files

The Access Graph feature has been incorporated in 'module.go', 'authservice.proto', 'auth.go', and 'plugin.go' files. A method 'EnableAccessGraph' is also introduced in 'module.go' to allow the usage of this feature. Descriptions of Access Graph have been adjusted across files to explicitly state the feature's enablement.

* Fix tests

* Remove unused access graph feature code

This commit removes all references and uses of the access graph feature from the codebase. It also updates corresponding tests to reflect this removal. It also adds EnableAccessGraph() method to testing module to fix test compilation.

* [TAG] Add localStorage value for enabling TAG & SQL editor (#34111)

* Add localStorage value for enabling TAG

* Add key for enabling the manual SQL editor

* Add access graph features retrieval in UserService

The code adds a `fetchAccessGraphFeatures` in UserService that retrieves and stores the access graph features from the provided API URL into the local storage. It also includes error handling to log the error and continue if fetching access graph features fails. The changes also touch the config file to include the API URL for fetching the access graph features.

---------

Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>

* Revert import removal - TS (#34713)

#34111 by mistake removed an import. That change broke the UI. This PR reverts that change.

* Update localStorage item key for AccessGraphSQLEnabled (#34722)

The localStorage item key corresponding to the 'AccessGraphSQLEnabled' state has been updated. This is a typo from the initial implementation.

---------

Co-authored-by: Ryan Clark <ryan.clark@goteleport.com>