fix: External Audit Storage: check for athena URI and ignore others (#34715)
* fix: eas: check for athena URI and ignore others

There must be an existing athena events URI for External Audit Storage
to be enabled.
If users enable External Audit Storage they do not want us to store
events on our infra, so this avoids setting up a MultiLog that would
write events to multiple backends if multiple audit URIs are passed.

* Apply suggestions from code review

Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>

* use package level error

---------

Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>