Pin additional 3rd party GitHub actions (#34678)
We favor determinism and security in our CI and CD automation.

As part of this, we want to ensure we can review, changes, roll them out
at an expected time, and roll them back if something breaks.  All of
this is enabled by using pinned versions of GitHub actions instead of
floating tags.

This PR pins all actions in this repo, except for:

* First party actions (from a `gravitational/` repo). We control the
  code review and deploy process for these.
* 2nd party actions (from the `actions/` or `github/` repos). We trust
  that if these were compromised, we'd not be the most valuable target.
  Furthermore, these are by far the most common actions, and we save a
  deal of sustaining work by not manually approving upgrades.

Contributes to https://github.com/gravitational/security-findings/issues/50