Rerun login hooks when a user's access has changed. (#2605)
When a user's access changes, login hooks are re-run for the affected users.
A user's access is considered changed when any of the following happens:

- A user resource changes in the backend.
- A user's membership to an access list is changed.
- An access list that a user belongs to is changed in a way that doesn't
  reflect the user login states stored in the backend.
- A role attached to a user or access list is changed.

This will ensure that, when users are removed from access lists or user
roles are changed, then the Okta user assignment creator will be re-run and
Okta access will be recalculated. This will actively update the Okta access
of users when user access is changed, meaning that Okta access will not
require a new login for a user's Okta access is revoked.

A small change was made to the Okta user assignment creator which allows for
user assignments with only groups or apps to be made. Tests were updated to
test for this.

Source-commit: gravitational/teleport.e@27b2dda5d17754a60ba619e5bf5d21d4c655f2de
Source-pull-requests: gravitational/teleport.e#2605