External Audit Storage: raise cluster alerts for elevated error rates
This commit adds a component that observes and counts errors relevent to
External Audit Storage, and periodically raises or clears cluster alerts
based on the recent error rate.

This information is being surfaced to users via cluster alerts because
External Audit Storage is specifically for Cloud users who have no way
to check auth server logs or metrics, and the errors may be caused by
their own error.

Possible error conditions this is meant to catch:
- deleted/modified OIDC connector
- deleted/modified IAM role
- deleted/modified IAM policy
- deleted S3 buckets
- deleted Athena workgroups
- deleted/modified Glue tables
- other possible permissions issues