Removes test `DeleteRole()` call from production `DeleteUser()` (#34108)
Prior to this patch, the production `auth` server appears to be trying to delete
a user-specific role before deleting the user itself. This user-specific role appears
to be only created as part of some test harness setup, and the concept should never have
leaked into the production `auth` server.

This patch removes the code in question, but does not isolate the code that leaked in
to the producton service. As the test code is used by multiple packages for their own
tests, any attempt to do would trigger a large refactor, that would swamp this functional
change.

Changelog: none