Connect My Computer: Reload certs if user already has role but role got updated with new login (#34879)
* Untangle assert on CertsReloaded from assigning existing role

We used to assume that if the CMC role already exists and it's assigned
to the user already, then CertsReloaded must be false.

This assumption is wrong. If the user already has the role but the role
gets updated because it didn't include the current system username,
the certs need to be reloaded in order to refresh the list of available
logins.

To be able to represent this scenario as a test case, we must untangle
the expectation on CertsReloaded from the act of adding the existing role
to the user.

* Reload certs if user already has role but logins in role got updated