Simplify `Dockerfile-centos7-fips` and prepare for FIPS arm64 builds
Replace `Dockerfile-centos7-fips` with a simple version that just uses
the existing `Dockerfile-centos7` image with some extra modifications
(specifically, enabling BoringCrypto for Go and configuring OpenSSL to
use FIPS mode).

Several changes were made to `Dockerfile-centos7` to support this.

There is a mix of `public.ecr.aws` and `ghcr.io` being used for the
buildboxes. The default now uses `ghcr.io`.

Buildbox images on ECR did not include the architecture as part of
the tag, while GHCR did. Dronegen is changed to include the
architecture in ECR as well for consistency.

Additionally, start building arm and CentOS 7 FIPS buildbox images
in GHA to bring it in sync with `dronegen/buildbox.go`, as well as
additionally build on amd64 for the two CentOS 7 buildboxes.

Ref #5068.
Ref #10581.